projects/flask-config/pyproject.toml
[project]
name = "flaskr"
version = "0.1"
requires-python = ">=3.13"
dependencies = ["flask", "python-dotenv"]

[tool.pyright]
venvPath = "."
venv = ".venv"
